install.php change/correcting Yiipath,System check
Including HTML meta charset UTF-8 show value if stored in the session path to yii, for if the user refreshes the page. 
Trim blanks to get $ _POST ['yiicheck] [' path '] 
Correcting variables that have not been previously initialized Changed the condition "or" with "and" to add the / final "yiiPath" Changing the return error / ok for ajax in steps "yiiPath" and "checkSystem":
1) Only error if a network failure. Check if there is a failure of PHP. 
2) The values returned are the JSON structure, which indicates whether the step has been done well or badly. 
3) Added the php functions "Sendok", "sendError", "echoSend". Unified time / code to check the system configuration. 
Shows the system requirements with explanatory text alongside. For a quick read. 
Added the Next button when the system check is ok.